Integrates spring-data config
This adds a spring-data Repository for OfferingEntity and
evaluates how it integrates into the DAO module. It adds
a few simple tests which show how to use arepository (a)
within tests and (b) within the SPI implementation layer.

We have a naming conflict with the current module which
uses the term 'Repository' already for sticking together the
domain objects from raw entities using plain DAOs. We will
have to resolve this by dropping the term 'Repository' and
replacing it by 'Assembler' to make the distinction clear
between a spring-data Repository and SPI domain object
assembly logic. This however requires further refactoring.