Init Spring data integration
This work starts to integrate with spring data. Integration
by means of getting rid of all the custom DAO classes and
replace them with Repositories (have to rename the former
Repository concept to Aggregate or similar).

Actual goal is better integration with Spring data to get
CRUD out-of-the-box, testability, common design patterns, etc.

This version also includes QueryDSL to provide a more DDD-like
structuring of SQL queries. The series-hibernate module provides
generated Q-code based on the mapping files available. The Q-code
can also be generated directly on a database level.