Rework SQL DSL specification API to work with DbQuery
Let QuerySpecification be creatable via static constructor method which
takes the DbQuery as an argument. This moves most of the expression
creation to the Specification itself and can be combined to simple query
expression like `matchFilter()`.

It also offers creation of Subqueries to match entities depending on 1:n
relationship without direct FK. This is done by subselect on the FK.

See example in ParameterOutputAssembler:

```
  DatasetQuerySpecifications dsFilterSpec =
DatasetQuerySpecifications.of(query);
  JPQLQuery<DatasetEntity> filteredDatasets =
dsFilterSpec.filteredDatasets();

  OfferingQuerySpecifications oFilterSpec =
OfferingQuerySpecifications.of(query);
  BooleanExpression predicate =
oFilterSpec.subSelectFrom(filteredDatasets);

  Iterable<E> findAll = parameterRepository.findAll(predicate);
```