Check against NotInitializedDatasetEntity value types
For some use cases a two step process is needed to end with
qualified datasets. For example the transactional SOS workflow
needs to be able to persist some sort of dataset templates which
 can become fully qualified only when real data gets associated
with it at a later step. For this reason, not initialized datasets
get filtered by default.

A repository Query will be used to qualify (not yet initialized)
datasets. It autoclears automatically persistence context, so
that entities within the current context can be retrieved in an
updated manner (i.e. entities which have been an instance of
NotInitializedDatasetEntity are returned qualified datasets by
the repository)