Let GeometryEntity set GeometryFactory when needed
Setting GeometryEntity seems to be better located at GeometryEntity
as it decides when it is actually needed to be set (when creating Point
from lat/lon values).

However, it is still weird to let an entity perform such logic. During
this broader refactoring phase, however, this move will make some
things a bit simpler in a later stage