Refactor assemble data
Change parameter from DatasetEntity to dataset id (long)to avoid conlficts with first and last observations when querying data/observations. If the dataset is already loaded in the session, the first and last observations are loaded as proxy values and occur in the result list of the observation query as proxy values which results in an error when processing the observations.